review update, also move sanitize fn to parser file
